Fix a warning. (Edgar Reynaldo)

+++ src/file.c
@@ -559,7 +559,7 @@ int al_vfprintf(ALLEGRO_FILE *pfile, const char *format, va_list args)
             if (size > 0)
             {
                rv = al_fwrite(pfile, (const void*)(al_cstr(ustr)), size);
-               if (rv != size) {
+               if (rv != (int)size) {
                   rv = -1;
                }
             }
